Migrate some global schedule queries endpoints
