+++ server/service/service_global_schedule.go
@@ -7,33 +7,6 @@ import (
 	"github.com/fleetdm/fleet/v4/server/ptr"
 )
 
-func (svc *Service) GlobalScheduleQuery(ctx context.Context, sq *fleet.ScheduledQuery) (*fleet.ScheduledQuery, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.Pack{}, fleet.ActionRead); err != nil {
-		return nil, err
-	}
-
-	gp, err := svc.ds.EnsureGlobalPack(ctx)
-	if err != nil {
-		return nil, err
-	}
-	sq.PackID = gp.ID
-
-	return svc.ScheduleQuery(ctx, sq)
-}
-
-func (svc *Service) GetGlobalScheduledQueries(ctx context.Context, opts fleet.ListOptions) ([]*fleet.ScheduledQuery, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.Pack{}, fleet.ActionRead); err != nil {
-		return nil, err
-	}
-
-	gp, err := svc.ds.EnsureGlobalPack(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	return svc.ds.ListScheduledQueriesInPack(ctx, gp.ID, opts)
-}
-
 func (svc *Service) ModifyGlobalScheduledQueries(
 	ctx context.Context,
 	id uint,
