+++ server/service/transport_global_schedule.go
@@ -6,16 +6,6 @@ import (
 	"net/http"
 )
 
-func decodeGetGlobalScheduleRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	opts, err := listOptionsFromRequest(r)
-	if err != nil {
-		return nil, err
-	}
-	var req getGlobalScheduleRequest
-	req.ListOptions = opts
-	return req, nil
-}
-
 func decodeModifyGlobalScheduleRequest(ctx context.Context, r *http.Request) (interface{}, error) {
 	id, err := uintFromRequest(r, "id")
 	if err != nil {
@@ -38,15 +28,5 @@ func decodeDeleteGlobalScheduleRequest(ctx context.Context, r *http.Request) (in
 	}
 	var req deleteGlobalScheduleRequest
 	req.ID = uint(id)
-	return req, nil
-}
-
-func decodeGlobalScheduleQueryRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	var req globalScheduleQueryRequest
-
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		return nil, err
-	}
-
 	return req, nil
 }
